Ensure install .desktop file is always removed.
Summary: Fixes T82.

Test Plan: Install and reboot. Desktop should be free of the install .desktop file. This has been tested already for EFI and BIOS in a virtual machine, but since this is related to a task that concerns a rather particular system consisting of 32 bit EFI with a 64 bit processor, it will need to be tested in other cases.

Reviewers: tsimonq2

Reviewed By: tsimonq2

Maniphest Tasks: T82

Differential Revision: https://phab.lubuntu.me/D30